wcloud/pyproject.toml
[project]
name = "wcloud"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "ifritJP", email = "15117169+ifritJP@users.noreply.github.com" }
]
dependencies = [
    "mecab-python3>=1.0.8",
    "unidic>=1.1.0",
    "wordcloud>=1.9.3",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true
